Added POST_Request_Ticker_Index and updated CAPM
POST_Request_Ticker_Index handles API grabs and JSON parsing. It will grab the adjusted daily closing price for any equity, etf, or index from the past 5 years.

CAPM was changed dramatically. It originally handled the API grabs and JSON parsing, but it was quickly becoming too complicated and convoluted. It now is split up to find the adjusted annual market return, the beta, and the expected return of an asset via the Capital Asset Pricing Model.

No changes to governance